Refactoring example

+++ oracle_example/src/main/java/com/github/braisdom/objsql/example/oracle/PersistenceExample.java
@@ -52,7 +52,7 @@ public class PersistenceExample {
         extendedAttributes.put("age", 28);
 
         Map<String, Object> attributes = new HashMap<>();
-        attributes.put("id", 10);
+        attributes.put("id", BigDecimal.valueOf(10));
         attributes.put("no", "200000");
         attributes.put("name", "Carter");
         attributes.put("gender", 1);
@@ -69,7 +69,7 @@ public class PersistenceExample {
         extendedAttributes.put("age", 28);
 
         Map<String, Object> attributes = new HashMap<>();
-        attributes.put("id", 9);
+        attributes.put("id", BigDecimal.valueOf(9));
         attributes.put("no", "200000");
         attributes.put("name", "Barbara\t");
         attributes.put("gender", 1);
